remove mutex on debouncer and layout, remove layout_register task
the layout_register task and the mutex on the layout and debouncer isn't really necessary.
removing the mutex required giving the layout_collect task ownership over the layout data, which should be fine.
removing the mutex also allowed for setup_keyboard_layout and setup_keyboard_matrix to be turned into functions instead of macros.